+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.6.7] - 2020-10-21
+
 ### Fixed
 
-- `SoftDeletes` methods on relations are no longer marked as undefined ([#692](https://github.com/nunomaduro/larastan/pull/692))
+- `SoftDeletes` methods on relations are no longer marked as undefined ([#692](https://github.com/nunomaduro/larastan/pull/692)) Thanks @jdrieghe
 - Generic model type is preserved when `with` method is used on a model instance.
 
 ## [0.6.6] - 2020-10-17
@@ -431,7 +433,8 @@ Upgrade guide: [UPGRADE.md](https://github.com/nunomaduro/larastan/blob/master/U
 ### Added
 - Adds first alpha version
 
-[Unreleased]: https://github.com/nunomaduro/larastan/compare/v0.6.6...HEAD
+[Unreleased]: https://github.com/nunomaduro/larastan/compare/v0.6.7...HEAD
+[0.6.7]: https://github.com/nunomaduro/larastan/compare/v0.6.6...v0.6.7
 [0.6.6]: https://github.com/nunomaduro/larastan/compare/v0.6.5...v0.6.6
 [0.6.5]: https://github.com/nunomaduro/larastan/compare/v0.6.4...v0.6.5
 [0.6.4]: https://github.com/nunomaduro/larastan/compare/v0.6.3...v0.6.4